+++ functions/index.ts
@@ -1,0 +1,1 @@
+export {rollupDailyMetrics} from './src/rollupDailyMetrics';